+++ satella/coding/sequences/__init__.py
@@ -3,7 +3,7 @@ from .choose import choose, choose_one
 from .iterators import infinite_counter, take_n, is_instance, skip_first, zip_shifted, \
     stop_after, iter_dict_of_list, shift, other_sequence_no_longer_than, count, even, \
     odd, n_th, smart_enumerate, smart_zip, unique, ConstruableIterator, walk, length, map_list, \
-    is_empty, IteratorListAdapter, enumerate2, to_iterator, ListWrapperIterator
+    is_empty, IteratorListAdapter, enumerate2, to_iterator, ListWrapperIterator, try_close
 from .sequences import is_last, add_next, half_cartesian, group_quantity, Multirun, make_list, \
     infinite_iterator, filter_out_false, filter_out_nones, index_of, index_of_max
 
@@ -14,4 +14,4 @@ __all__ = ['choose', 'choose_one', 'infinite_counter', 'take_n', 'is_instance', 
            'even', 'odd', 'Multirun', 'smart_enumerate', 'unique', 'map_list', 'is_empty',
            'make_list', 'RollingArithmeticAverage', 'index_of', 'index_of_max',
            'IteratorListAdapter', 'enumerate2', 'infinite_iterator', 'to_iterator',
-           'filter_out_false', 'filter_out_nones', 'ListWrapperIterator']
+           'filter_out_false', 'filter_out_nones', 'ListWrapperIterator', 'try_close']