+++ tests/test_coding/test_sequences.py
@@ -6,11 +6,19 @@ from satella.coding.sequences import choose, choose_one, infinite_counter, take_
     iter_dict_of_list, shift, other_sequence_no_longer_than, count, even, odd, Multirun, n_th, \
     unique, length, map_list, smart_zip, is_empty, make_list, RollingArithmeticAverage, \
     enumerate2, infinite_iterator, to_iterator, filter_out_false, filter_out_nones, \
-    index_of, index_of_max
+    index_of, index_of_max, try_close
 from satella.coding.predicates import x
 
 
 class TestSequences(unittest.TestCase):
+
+    def test_try_close(self):
+        def iterator():
+            yield 2
+
+        a = iterator()
+        try_close(a)
+        try_close(iter([1, 2, 3]))
 
     def test_index_of_max(self):
         self.assertRaises(ValueError, lambda: index_of_max([]))