+++ satella/coding/structures/__init__.py
@@ -16,7 +16,10 @@ from .sorted_list import SortedList, SliceableDeque
 from .sparse_matrix import SparseMatrix
 from .typednamedtuple import typednamedtuple
 from .lru import LRU
+from .syncable_droppable import DBStorage, SyncableDroppable
+
 __all__ = [
+    'DBStorage', 'SyncableDroppable',
     'LRU',
     'LRUCacheDict',
     'HashableMixin',