+++ satella/instrumentation/__init__.py
@@ -1,5 +1,6 @@
 from .dump_frames_on import install_dump_frames_on
-from .trace_back import Traceback, GenerationPolicy, StackFrame, StoredVariableValue
+from .trace_back import Traceback, GenerationPolicy, StackFrame, StoredVariableValue, \
+    frame_from_traceback
 
 __all__ = ['install_dump_frames_on', 'Traceback', 'GenerationPolicy', 'StoredVariableValue',
-           'StackFrame']
+           'StackFrame', 'frame_from_traceback']