+++ satella/instrumentation/trace_back/__init__.py
@@ -1,4 +1,5 @@
 from .classes import GenerationPolicy, StackFrame, StoredVariableValue
-from .trace_back import Traceback
+from .trace_back import Traceback, frame_from_traceback
 
-__all__ = ['GenerationPolicy', 'StoredVariableValue', 'StackFrame', 'Traceback']
+__all__ = ['GenerationPolicy', 'StoredVariableValue', 'StackFrame', 'Traceback',
+           'frame_from_traceback']