+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 # v2.14.9
 
+* added `KVTuple`
+* fixed `Appendable`
+* added `Comparable`