+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.14.10
+
+* added `SyncableDroppable.get_last_value`
+* added `SyncableDroppable.get_archive`