+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.14.11
+
+* added `frame_from_traceback`