+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.14.11
-
-* added `frame_from_traceback`
+# v2.14.12